Adding initial structure for ITKv4 External Modules
Adds the basic cmake structures needed to create a ITKv4 External
module. The intention is that this repository can be merged into a
new one to be a starting point to begin developmnet. The
ExternalTemplate string will need to be replaced with the name of the
new module, and some files removed.